Fix budget precision loss by changing FLOAT to DOUBLE columns
MySQL FLOAT (32-bit) silently rounds values at ~7 significant digits,
causing budget values like 1051785 to be stored as 1051780. DOUBLE
(64-bit, ~15 digits) preserves all values exactly.

Includes pre-deploy migration SQL: migrate-float-to-double.sql

+++ back-end/src/entities/porb-bilateral.entity.ts
@@ -48,7 +48,7 @@ export class PorbBilateral {
   bilateral_outputs: string;
 
   @ApiProperty()
-  @Column({ type: 'float', nullable: true })
+  @Column({ type: 'double', nullable: true })
   bilateral_budget: number;
 
   @ApiProperty()
